+++ amethyst_animation/src/resources.rs
@@ -719,13 +719,14 @@ where
     ///
     /// This should be used with care, as this will leave all linked samplers in place. If in
     /// doubt, use `abort()` instead.
-    pub fn remove(&mut self, id: I) {
+    pub fn remove(&mut self, id: I) -> &mut Self {
         if let Some(index) = self.animations.iter().position(|a| a.0 == id) {
             self.animations.remove(index);
         }
+        self
     }
 
-    fn set_command(&mut self, id: I, command: AnimationCommand<T>) {
+    fn set_command(&mut self, id: I, command: AnimationCommand<T>) -> &mut Self {
         if let Some(&mut (_, ref mut control)) = self.animations.iter_mut().find(|a| a.0 == id) {
             control.command = command;
         } else if let Some(ref mut control) = self
@@ -735,20 +736,22 @@ where
         {
             control.control.command = command;
         }
+
+        self
     }
 
     /// Start animation if it exists
-    pub fn start(&mut self, id: I) {
-        self.set_command(id, AnimationCommand::Start);
+    pub fn start(&mut self, id: I) -> &mut Self {
+        self.set_command(id, AnimationCommand::Start)
     }
 
     /// Pause animation if it exists
-    pub fn pause(&mut self, id: I) {
-        self.set_command(id, AnimationCommand::Pause);
+    pub fn pause(&mut self, id: I) -> &mut Self {
+        self.set_command(id, AnimationCommand::Pause)
     }
 
     /// Toggle animation if it exists
-    pub fn toggle(&mut self, id: I) {
+    pub fn toggle(&mut self, id: I) -> &mut Self {
         if let Some(&mut (_, ref mut control)) = self.animations.iter_mut().find(|a| a.0 == id) {
             if control.state.is_running() {
                 control.command = AnimationCommand::Pause;
@@ -756,10 +759,12 @@ where
                 control.command = AnimationCommand::Start;
             }
         }
+
+        self
     }
 
     /// Set animation rate
-    pub fn set_rate(&mut self, id: I, rate_multiplier: f32) {
+    pub fn set_rate(&mut self, id: I, rate_multiplier: f32) -> &mut Self {
         if let Some(&mut (_, ref mut control)) = self.animations.iter_mut().find(|a| a.0 == id) {
             control.rate_multiplier = rate_multiplier;
         }
@@ -770,26 +775,28 @@ where
         {
             control.control.rate_multiplier = rate_multiplier;
         }
+
+        self
     }
 
     /// Step animation
-    pub fn step(&mut self, id: I, direction: StepDirection) {
-        self.set_command(id, AnimationCommand::Step(direction));
+    pub fn step(&mut self, id: I, direction: StepDirection) -> &mut Self {
+        self.set_command(id, AnimationCommand::Step(direction))
     }
 
     /// Set animation input value (point of interpolation)
-    pub fn set_input(&mut self, id: I, input: f32) {
-        self.set_command(id, AnimationCommand::SetInputValue(input));
+    pub fn set_input(&mut self, id: I, input: f32) -> &mut Self {
+        self.set_command(id, AnimationCommand::SetInputValue(input))
     }
 
     /// Set blend weights
-    pub fn set_blend_weight(&mut self, id: I, weights: Vec<(usize, T::Channel, f32)>) {
-        self.set_command(id, AnimationCommand::SetBlendWeights(weights));
+    pub fn set_blend_weight(&mut self, id: I, weights: Vec<(usize, T::Channel, f32)>) -> &mut Self {
+        self.set_command(id, AnimationCommand::SetBlendWeights(weights))
     }
 
     /// Abort animation
-    pub fn abort(&mut self, id: I) {
-        self.set_command(id, AnimationCommand::Abort);
+    pub fn abort(&mut self, id: I) -> &mut Self {
+        self.set_command(id, AnimationCommand::Abort)
     }
 
     /// Add animation with the given id, unless it already exists
@@ -800,20 +807,20 @@ where
         end: EndControl,
         rate_multiplier: f32,
         command: AnimationCommand<T>,
-    ) {
-        if self.animations.iter().any(|a| a.0 == id) {
-            return;
+    ) -> &mut Self {
+        if !self.animations.iter().any(|a| a.0 == id) {
+            self.animations.push((
+                id,
+                AnimationControl::new(
+                    animation.clone(),
+                    end,
+                    ControlState::Requested,
+                    command,
+                    rate_multiplier,
+                ),
+            ));
         }
-        self.animations.push((
-            id,
-            AnimationControl::new(
-                animation.clone(),
-                end,
-                ControlState::Requested,
-                command,
-                rate_multiplier,
-            ),
-        ));
+        self
     }
 
     /// Add deferred animation with the given id, unless it already exists
@@ -826,29 +833,29 @@ where
         command: AnimationCommand<T>,
         wait_for: I,
         wait_deferred_for: DeferStartRelation,
-    ) {
-        if self.animations.iter().any(|a| a.0 == id) {
-            return;
+    ) -> &mut Self {
+        if !self.animations.iter().any(|a| a.0 == id) {
+            self.deferred_animations.push(DeferredStart {
+                animation_id: id,
+                relation: (wait_for, wait_deferred_for),
+                control: AnimationControl::new(
+                    animation.clone(),
+                    end,
+                    ControlState::Requested,
+                    command,
+                    rate_multiplier,
+                ),
+            });
         }
-        self.deferred_animations.push(DeferredStart {
-            animation_id: id,
-            relation: (wait_for, wait_deferred_for),
-            control: AnimationControl::new(
-                animation.clone(),
-                end,
-                ControlState::Requested,
-                command,
-                rate_multiplier,
-            ),
-        });
+        self
     }
 
     /// Insert an animation directly
-    pub fn insert(&mut self, id: I, control: AnimationControl<T>) {
-        if self.animations.iter().any(|a| a.0 == id) {
-            return;
+    pub fn insert(&mut self, id: I, control: AnimationControl<T>) -> &mut Self {
+        if !self.animations.iter().any(|a| a.0 == id) {
+            self.animations.push((id, control));
         }
-        self.animations.push((id, control));
+        self
     }
 
     /// Check if there is an animation with the given id in the set
@@ -905,8 +912,9 @@ where
     }
 
     /// Insert an animation in the set
-    pub fn insert(&mut self, id: I, handle: Handle<Animation<T>>) {
+    pub fn insert(&mut self, id: I, handle: Handle<Animation<T>>) -> &mut Self {
         self.animations.insert(id, handle);
+        self
     }
 
     /// Retrieve an animation handle from the set