+++ amethyst_input/src/util.rs
@@ -52,3 +52,25 @@ pub fn get_input_axis_simple<T: BindingTypes>(
         .and_then(|ref n| input.axis_value(n))
         .unwrap_or(0.0)
 }
+
+/// If this event was for manipulating a mouse button, this will return the `MouseButton`
+/// and the new state.
+pub fn get_mouse_button(event: &Event) -> Option<(MouseButton, ElementState)> {
+    match *event {
+        Event::WindowEvent { ref event, .. } => match *event {
+            WindowEvent::MouseInput { button, state, .. } => Some((button, state)),
+            _ => None,
+        },
+        _ => None,
+    }
+}
+
+/// Returns true if the event passed in is a mouse button down event for the
+/// provided `MouseButton`.
+pub fn is_mouse_button_down(event: &Event, button: MouseButton) -> bool {
+    if let Some((pressed_button, state)) = get_mouse_button(event) {
+        return pressed_button == button && state == ElementState::Pressed;
+    } else {
+        return false;
+    }
+}